Switch input spot form pages available
