+++ frontend/src/components/Spot/CreateSpotForm.scss
@@ -118,6 +118,7 @@
             height: 150px;
             font-size: 18px;
             font-weight: 600;
+            margin-top: 30px;
 
             textarea {
                 height: 100px;
@@ -183,6 +184,12 @@
         }
         // background-color: red;
     }
+
+    // .prevSubmitButton {
+    //     display: flex;
+    //     // justify-content: center;
+    //     // align-content: center;
+    // }
 }
